Changed Idempotent

+++ TaskManagement.Core.Application/Features/Commands/Issue/DeleteIssueCommand.cs
@@ -1,5 +1,8 @@
-﻿using MediatR;
+﻿using FluentValidation;
+using MediatR;
+using System.Security.Claims;
 using TaskManagement.Core.Application.Exceptions;
+using TaskManagement.Core.Application.ExtensionMethods;
 using TaskManagement.Core.Application.Interfaces;
 
 namespace TaskManagement.Core.Application.Features.Commands.Issue
@@ -7,22 +10,33 @@ namespace TaskManagement.Core.Application.Features.Commands.Issue
     public class DeleteIssueCommand : IRequest<int>
     {
         public int IssueId { get; }
-        public DeleteIssueCommand(int issueId)
+        public ClaimsPrincipal User { get; }
+        public DeleteIssueCommand(int issueId, ClaimsPrincipal user)
         {
             IssueId = issueId;
+            User = user;
         }
         public class DeleteIssueCommandHandler : IRequestHandler<DeleteIssueCommand, int>
         {
             private readonly IIssueRepository _issueRepository;
-            public DeleteIssueCommandHandler(IIssueRepository issueRepository)
+            private readonly IUserRepository _userRepository;
+            public DeleteIssueCommandHandler(IIssueRepository issueRepository, IUserRepository userRepository)
             {
                 _issueRepository = issueRepository;
+                _userRepository = userRepository;
             }
             public async Task<int> Handle(DeleteIssueCommand request, CancellationToken cancellationToken)
             {
                 var issue = await _issueRepository.GetAsync(request.IssueId);
                 if (issue == null)
                     throw new EntityNotFoundException();
+
+                var currentUser = await request.User.MapToDatabase(_userRepository);
+                var currentUserIsAdmin = currentUser.Roles.Select(x => x.Role).Any(x => x.IsAdmin);
+                
+                if (!currentUserIsAdmin && currentUser.Id != issue.ReporterId && currentUser.Id != issue.AssigneeId)
+                    throw new ValidationException("Access is denied");
+
                 var result = await _issueRepository.DeleteAsync(request.IssueId);
                 return result;
             }
